refactor: styles files
