--- lib/styles/colors.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-// aimimi theme color palette
-const themeColor = Color(0xff80D3DC);
-const themeShadedColor = Color(0xff1C4B56);
-const themeMutedColor = Color(0xff759198);
-
-// mono color palette
-const monoPrimaryColor = Color(0xff000000);
-const monoSecondaryColor = Color(0xff999999);
-const monoTintedColor = Color(0xffC2C2C2);
-
-// background color palette
-const backgroundColor = Color(0xffF2F4F6);
-const backgroundTintedColor = Color(0xffF8FAFB);
-const backgroundShadedColor = Color(0xffEFF0F1);
-
-// default overlay color
-// const overlayColor = Color(0xff737373);
