--- lib/styles/text_fields.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-const textFieldBorder = OutlineInputBorder(
-  borderSide: BorderSide(
-    width: 0,
-    style: BorderStyle.none,
-  ),
-  borderRadius: BorderRadius.all(Radius.circular(8)),
-);
