--- lib/styles/text_styles.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'colors.dart';
-
-const appBarTitleTextStyle = TextStyle(
-  color: themeShadedColor,
-  fontSize: 22,
-  fontWeight: FontWeight.w700,
-);
-
-const textFieldTitleTextStyle = TextStyle(
-  color: themeShadedColor,
-  fontSize: 16,
-  fontWeight: FontWeight.w700,
-);
-
-const textFieldHintTextStyle = TextStyle(
-  fontSize: 15,
-  fontWeight: FontWeight.w400,
-  color: monoSecondaryColor,
-);
