+++ lib/views/main_view.dart
@@ -1,5 +1,4 @@
-import 'package:aimimi/styles/colors.dart';
-import 'package:aimimi/styles/text_styles.dart';
+import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/views/today_view.dart';
 import 'package:aimimi/widgets/modal/modal_add_goal.dart';
 import 'package:aimimi/views/leaderboard_view.dart';
