+++ lib/widgets/goal/goal.dart
@@ -1,4 +1,4 @@
-import 'package:aimimi/styles/colors.dart';
+import 'package:aimimi/constants/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
